refactor(LineSectionExplorer): explore line without side effects
- Line의 과도하게 커진 책임을 분산하기 위해 Line 내부 상태를 탐색하는 기능 중 일부를 별도의 클래스로 분리
- Sections를 더하고 빼는 기능은 Line내의 객체 레퍼런스를 다른 객체에서 고치는 상태가 일어나기 때문에 side effect 추적이 어렵다 판단되어 보류
- 조회의 경우 복사본을 사용하면 side effect로부터 자유롭게 탐색이 가능하기 때문에 조회 기능을 따로 분리
