+++ src/test/java/nextstep/subway/line/domain/LineTest.java
@@ -18,31 +18,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class LineTest {
-    @DisplayName("등록된 구간의 상행종점역을 찾을 수 있다.")
-    @Test
-    void findFirstStationTest() {
-        String name = "2호선";
-        String color = "초록색";
-        int distance = 3;
-        Line line = new Line(name, color, StationFixtures.강남역, StationFixtures.잠실역, distance);
-
-        Station firstStation =  line.findUpStation();
-
-        assertThat(firstStation.getName()).isEqualTo(StationFixtures.강남역.getName());
-    }
-
-    @DisplayName("등록된 구간이 없는 Line의 상행종점역을 찾으면 예외 발생")
-    @Test
-    void findFirstStationFailTest() {
-        String name = "2호선";
-        String color = "초록색";
-        Line line = new Line(name, color);
-
-        assertThatThrownBy(line::findUpStation)
-                .isInstanceOf(ExploreSectionException.class)
-                .hasMessage("해당 노선의 첫번째 구간을 찾을 수 없습니다.");
-    }
-
     @DisplayName("구간 순서대로 정렬된 역 목록을 받을 수 있다.")
     @Test
     void getStationsTest() {
